+++ custom/justech_l10n_do_ncf/models/__init__.py
@@ -13,5 +13,6 @@ from . import ncf_consumption
 from . import ncf_admin_center
 from . import ncf_migration_log
 from . import purchase_emission_config
+from . import l10n_latam_document_type
 from . import account_move
 from . import sale_order